find_command(): move from avocado.utils.process to avocado.utils.path
The reasons for the proposed move include:

1) Even though find_command() looks for executables, it doesn't really
   try to execute them as a living process.

2) The implementation looks for it in a $PATH like list, so, out of a
   possibly better choice, avocado.utils.path was chosen.

3) Other modules can use that functionality, but not the execution of
   process.

4) Avoid circular dependencies: this is actually the first motivation
   of the proposed change, and a true story with the GDB plugin/test.
   Since avocado.utils.process depends on avocado.gdb, this module
   can not use find_command() to look for gdb/gdbserver binaries.
